chapter-10-deployment-pipelines/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
anthropic==0.18.1
openai==1.12.0
langchain==0.1.6
langchain-anthropic==0.1.1

# Vector Databases
pinecone-client==3.0.3
weaviate-client==4.4.0

# API Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# Kubernetes
kubernetes==29.0.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0